+++ src/fixeduint.rs
@@ -35,7 +35,8 @@ mod ilog_impl;
 mod isqrt_impl;
 mod iter_impl;
 mod midpoint_impl;
-mod mul_acc_ops_impl;
+#[cfg(feature = "modmath")]
+mod cios_row_ops_impl;
 mod mul_div_impl;
 mod multiple_impl;
 mod num_integer_impl;